Delete unnecessary .husky directory
The .husky directory carried over from the Prettier setup when I
branched from `set-up` (not main, now rebased).  It wasn't tracked by
git, so made it into this branch sneakily when I rebased this branch
onto main.

(Husky is the git hook manager we're using, implemented on the `set-up`
branch, to be submitted in a separate PR)