Move schemacrawler to separate docker-compose
docker-compose.yml stands up backing services.

Schemacrawler is a tool to create a database schema diagram, not a
backing service.  It also requires both the postgres backing service
to be stood up, as well as migrations to have been performed by Spring.

Therefore, there's a cyclic dependency: Schemacrawler requires
Postgres to be up and the Service to have run migrations.  Yet the
Service requires Postgres to be up in order to run migrations.
Therefore. Postgres and Schemacrawler cannot exist in the same
docker-compose.

This commit separates Schemacrawler into a separate docker-compose file
that can be run by the developer when the schema diagram needs to be
updated.