Configure all batch jobs to run asynchronously
The only batch job we currently have is triggered via an HTTP request.

We want the controller that triggers this job to be able to return a
response to the caller immediately, without waiting for a long-running
job to complete.

This config will launch jobs in a new thread and not block
immediate HTTP responses.